LayoutTests/platform/wk2/webarchive/loading/video-in-webarchive-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - willPerformClientRedirectToURL: resources/video-in-webarchive.webarchive
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
